Adding a uuid v4 implementation with crypto:rand_bytes/1
The default version uses random:uniform/1, which is based
on a broken implementation of the whitchman-hill PRNG. It
has bad divergence, which makes the UUID somewhat easy to guess.
On the other hand, crypto module's PRNG is crypto-safe and
shouldn't be as easy to guess, making a potential safe use
of the UUID for sessions or other values.